Intégration de google oauth et redesign de l'api

+++ models/sinistre.js
@@ -2,8 +2,8 @@ module.exports = (sequelize, Sequelize) => {
 
     const Sinistre = sequelize.define('Sinistre', {
         typeSin: { type: Sequelize.STRING, allowNull: false },
-        lon: { type: Sequelize.FLOAT, allowNull: false },
-        lat: { type: Sequelize.FLOAT, allowNull: false },
+        lon: { type: Sequelize.DOUBLE, allowNull: false },
+        lat: { type: Sequelize.DOUBLE, allowNull: false },
         image: { type: Sequelize.STRING, allowNull: false },
         dateSin: { type: Sequelize.DATE, defaultValue: Sequelize.NOW },
     }, { timestamps: false }, { underscored: true });
